chore: Enable capturing metrics at higher freq
