+++ app/server/appsmith-server/src/main/java/com/appsmith/server/configurations/MetricsConfig.java
@@ -1,0 +1,109 @@
+package com.appsmith.server.configurations;
+
+import com.appsmith.server.annotations.ConditionalOnMicrometerMetricsEnabled;
+import io.micrometer.core.aop.TimedAspect;
+import io.micrometer.core.instrument.MeterRegistry;
+import io.micrometer.observation.ObservationRegistry;
+import io.micrometer.observation.aop.ObservedAspect;
+import io.opentelemetry.api.OpenTelemetry;
+import io.opentelemetry.exporter.otlp.metrics.OtlpGrpcMetricExporter;
+import io.opentelemetry.instrumentation.micrometer.v1_5.OpenTelemetryMeterRegistry;
+import io.opentelemetry.sdk.OpenTelemetrySdk;
+import io.opentelemetry.sdk.metrics.Aggregation;
+import io.opentelemetry.sdk.metrics.InstrumentType;
+import io.opentelemetry.sdk.metrics.SdkMeterProvider;
+import io.opentelemetry.sdk.metrics.export.AggregationTemporalitySelector;
+import io.opentelemetry.sdk.metrics.export.DefaultAggregationSelector;
+import io.opentelemetry.sdk.metrics.export.PeriodicMetricReader;
+import io.opentelemetry.sdk.resources.Resource;
+import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnExpression;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+import java.time.Duration;
+
+/**
+ * This configuration file configures beans as required by NewRelic
+ */
+@RequiredArgsConstructor
+@Configuration
+public class MetricsConfig {
+
+    public static final String NEW_RELIC_MICROMETER_METRICS_ENDPOINT = "https://otlp.nr-data.net:4317";
+    public static final String API_KEY = "api-key";
+    public static final String SERVICE_NAME_KEY = "service.name";
+    public static final String INSTRUMENTATION_PROVIDER_KEY = "instrumentation.provider";
+    public static final String MICROMETER = "micrometer";
+
+    @Value("${appsmith.newrelic.micrometer.metrics.application.name}")
+    private String newRelicApplicationName;
+
+    @Value("${appsmith.newrelic.licensekey}")
+    private String newRelicKey;
+
+    @Value("${appsmith.newrelic.micrometer.metrics.container.name}")
+    private String newRelicContainerName;
+
+    private final CommonConfig commonConfig;
+
+    private int getMetricsInterval() {
+        if (commonConfig.isMetricsDetail()) {
+            return 1;
+        }
+        return 60;
+    }
+
+    @Bean
+    @ConditionalOnMicrometerMetricsEnabled
+    public TimedAspect timedAspect(MeterRegistry registry) {
+        return new TimedAspect(registry);
+    }
+
+    @Bean
+    @ConditionalOnMicrometerMetricsEnabled
+    public MeterRegistry meterRegistry(OpenTelemetry openTelemetry) {
+        return OpenTelemetryMeterRegistry.builder(openTelemetry).build();
+    }
+
+    @Bean
+    @ConditionalOnMicrometerMetricsEnabled
+    public OpenTelemetry openTelemetry() {
+        return OpenTelemetrySdk.builder()
+                .setMeterProvider(SdkMeterProvider.builder()
+                        .setResource(Resource.getDefault().toBuilder()
+                                .put(SERVICE_NAME_KEY, newRelicApplicationName)
+                                // Include instrumentation.provider=micrometer to enable micrometer metrics
+                                // experience in New Relic
+                                .put(INSTRUMENTATION_PROVIDER_KEY, MICROMETER)
+                                .put("container.name", newRelicContainerName)
+                                .build())
+                        .registerMetricReader(PeriodicMetricReader.builder(OtlpGrpcMetricExporter.builder()
+                                        .setEndpoint(NEW_RELIC_MICROMETER_METRICS_ENDPOINT)
+                                        .addHeader(API_KEY, newRelicKey)
+                                        // IMPORTANT: New Relic requires metrics to be delta temporality
+
+                                        .setAggregationTemporalitySelector(
+                                                AggregationTemporalitySelector.deltaPreferred())
+                                        // Use exponential histogram aggregation for histogram instruments
+                                        // to
+                                        // produce better data and compression
+                                        .setDefaultAggregationSelector(DefaultAggregationSelector.getDefault()
+                                                .with(
+                                                        InstrumentType.HISTOGRAM,
+                                                        Aggregation.base2ExponentialBucketHistogram()))
+                                        .build())
+                                // Match default micrometer collection interval of 60 seconds
+                                .setInterval(Duration.ofSeconds(getMetricsInterval()))
+                                .build())
+                        .build())
+                .build();
+    }
+
+    @Bean
+    @ConditionalOnExpression("${logging.verbose.enabled}")
+    ObservedAspect observedAspect(ObservationRegistry observationRegistry) {
+        return new ObservedAspect(observationRegistry);
+    }
+}
